feat: resize video, change pitch/speed (#7)
* feat: resize video module

* feat: speed/pitch module

* Update README

* fix: update linting, remove unused variable

* chore: add plausible env to .env.example, update README

* feat: update and format speed component

* feat: update and format resize component

* fix: retry trims that fail

* chore: remove comment

---------

+++ src/routes/VideoEditor.svelte
@@ -4,6 +4,7 @@
 	import ms from 'pretty-ms';
 	import { createEventDispatcher } from 'svelte';
 	import playIcon from '@iconify-icons/mdi/play-arrow';
+	import speedIcon from '@iconify-icons/mdi/speedometer-slow';
 	import pauseIcon from '@iconify-icons/mdi/pause';
 	import back10Icon from '@iconify-icons/mdi/rewind-10';
 	import forward10Icon from '@iconify-icons/mdi/fast-forward-10';
@@ -12,6 +13,7 @@
 	import convertIcon from '@iconify-icons/mdi/file-arrow-left-right';
 	import trimIcon from '@iconify-icons/mdi/content-cut';
 	import volumeIcon from '@iconify-icons/mdi/volume-high';
+	import resizeIcon from '@iconify-icons/mdi/aspect-ratio';
 	import compressIcon from '@iconify-icons/mdi/zip-box';
 	import bitrateIcon from '@iconify-icons/mdi/music-note';
 	import cropIcon from '@iconify-icons/mdi/crop';
@@ -19,7 +21,7 @@
 	import type { IconifyIcon } from '@iconify/svelte';
 	import PreviewStillContainer from './PreviewStillContainer.svelte';
 	import { MS_OPTIONS } from '$lib/util';
-	import { ffmpeg } from '$lib/ffmpeg';
+	import { ffmpeg, ffmpegAborted, runFFmpeg } from '$lib/ffmpeg';
 	import Trim from '$lib/components/video/Trim.svelte';
 	import Volume from '$lib/components/common/Volume.svelte';
 	import EditorTabs from '$lib/components/EditorTabs.svelte';
@@ -31,6 +33,9 @@
 	import Crop from '$lib/components/video/Crop.svelte';
 	import CropHandler from './CropHandler.svelte';
 	import ScreenshotButtons from './ScreenshotButtons.svelte';
+	import Resize from '$lib/components/video/Resize.svelte';
+	import Speed from '$lib/components/common/Speed.svelte';
+	import { generatePitchSpeedAudioCommand } from '$lib/utils/speed-utils';
 
 	export let dispatch = createEventDispatcher();
 	export let file: File | RemoteFile;
@@ -42,11 +47,6 @@
 	let modalOpen = false;
 	let processState = ProcessingState.IDLE;
 	let resultInfo: { elapsed: number; size: number } | null = null;
-
-	async function runFFmpeg(args: string[]) {
-		console.log(`Running command: ffmpeg ${args.join(' ')}`);
-		await ffmpeg.exec(args);
-	}
 
 	async function saveVideo() {
 		processState = ProcessingState.WRITING;
@@ -74,10 +74,47 @@
 			processState = ProcessingState.RUNNING;
 			const start = Date.now();
 			console.log(' ---- RUNNING FFmpeg ---- ');
-
+			const resizingVideo: boolean = ((!!resizeHeight) || (!!resizeWidth));
 			const converting = !!toExtension;
 			const bitrateChanged = bitrate > 0;
-			const otherFiltersUsed = volume !== 1 || volumeMode !== 0 || bitrateChanged || converting || willCrop || compressionLevel !== 0;
+
+			const speedFactorInternal: number = speedFactor ?? 1;
+
+			let rw2 = resizeWidth;
+			let rh2 = resizeHeight;
+
+			if (resizingVideo) {
+				const aspectRatio = willCrop ? Math.max(cropWidth,1)/Math.max(cropHeight,1) : Math.max(videoWidth,1)/Math.max(videoHeight,1);
+				if (!rw2) {
+					rw2 = rh2 * aspectRatio;
+				}
+				if (!rh2) {
+					rh2 = rw2 * (1 / aspectRatio);
+				}
+			}
+
+			const complexSpeedVideoCommand = speedFactorInternal !== 1 ? `setpts=PTS/${speedFactorInternal}` : '';
+			const complexCropVideoCommand = willCrop ? `crop=${cropWidth}:${cropHeight}:${cropX}:${cropY}` : '';
+			const complexResizeVideoCommand = resizingVideo ? `scale=ceil(${rw2}/2)*2:ceil(${rh2}/2)*2,setsar=1` : '';
+
+			const volumeFilter = (volume !== 1 || volumeMode !== 0) && volumeMode === 0 ? `volume=${volume.toFixed(2)}` : '';
+			const loudNormFilter = (volume !== 1 || volumeMode !== 0) && volumeMode === 1 ? `loudnorm=I=${loudnormArgs[0].toFixed(2)}:LRA=${loudnormArgs[1].toFixed(2)}:TP=${loudnormArgs[2].toFixed(2)}` : '';
+
+			const complexPitchSpeedAudioCommand = generatePitchSpeedAudioCommand(speedFactorInternal, keepPitch, semitoneFactor);
+
+			const complexVideoFilterPart = [complexCropVideoCommand, complexResizeVideoCommand, complexSpeedVideoCommand].filter(v => !!v);
+			const complexAudioFilterPart = [volumeFilter, loudNormFilter, complexPitchSpeedAudioCommand].filter(v=>!!v);
+			const allComplexFilters: string[] = [];
+			if(complexVideoFilterPart.length >= 1) {
+					allComplexFilters.push(`[0:v]${complexVideoFilterPart.join(',')}[v]`);
+			}
+
+			if(complexAudioFilterPart.length >= 1) {
+				allComplexFilters.push(`[0:a]${complexAudioFilterPart.join(',')}[a]`);
+			}
+
+
+			const otherFiltersUsed = volume !== 1 || volumeMode !== 0 || bitrateChanged || converting || willCrop || compressionLevel !== 0 || resizingVideo || allComplexFilters.length >= 1;
 			let trimOnNextCall = false;
 			const trimArgs = [
 				'-ss', ms(trimStart * 1000, MS_OPTIONS),
@@ -89,45 +126,64 @@
 
 			// For resource intensive calls, we trim first, then run other filters
 			if (willBeTrimmed && !trimOnNextCall) {
-				await runFFmpeg([
-					'-i',
-					`in.${extension}`,
-					...trimArgs,
-					...(trimReencoding ? ['-preset', 'ultrafast'] : ['-c:v', 'copy', '-c:a', 'copy']),
-					`clip.${extension}`
-				]);
-				await ffmpeg.deleteFile(`in.${extension}`);
-				await ffmpeg.rename(`clip.${extension}`, `in.${extension}`);
+				// We try two trimming attempts, if the first one fails, then force re-encoding on the next run
+				for (let attempt = 0; attempt < 2; attempt++) {
+					const forceReencoding = attempt !== 0;
+					console.info("Running intermediary FFmpeg command", forceReencoding ? '(forced re-encoding)' : '(initial run)');
+					await runFFmpeg([
+						'-i',
+						`in.${extension}`,
+						...trimArgs,
+						...((trimReencoding || forceReencoding) ? ['-preset', 'ultrafast'] : ['-c:v', 'copy', '-c:a', 'copy']),
+						`clip.${extension}`
+					]);
+					if ($ffmpegAborted && attempt === 0) {
+						console.warn('Failed to trim cleanly! Forcing re-encoding and trying again...');
+						continue;
+					}
+					await ffmpeg.deleteFile(`in.${extension}`);
+					await ffmpeg.rename(`clip.${extension}`, `in.${extension}`);
+					break;
+				}
 			}
 
-			if (!otherFiltersUsed) await ffmpeg.rename(`in.${extension}`, `out.${extension}`);
-			else
-				await runFFmpeg([
+
+			if (!otherFiltersUsed) {
+				// no other filters used
+				await ffmpeg.rename(`in.${extension}`, `out.${extension}`)
+			}
+			else {
+				const complexFilter: string = allComplexFilters.length >= 1 ? allComplexFilters.join(';') : '';
+
+
+				const ffCommand = [
 					'-i',
 					`in.${extension}`,
 					...(willBeTrimmed && trimOnNextCall ? trimArgs : []),
 					...(volume === 0
 						? ['-an']
-						: (volume !== 1 || volumeMode !== 0)
-							? volumeMode === 0
-								? ['-af', `volume=${volume.toFixed(2)}`]
-								: ['-af', `loudnorm=I=${loudnormArgs[0].toFixed(2)}:LRA=${loudnormArgs[1].toFixed(2)}:TP=${loudnormArgs[2].toFixed(2)}`]
-							: !converting && !bitrateChanged
+						: ((complexAudioFilterPart.length >= 1)
+							? []
+							: (!converting && !bitrateChanged && allComplexFilters.length === 0
 								? ['-c:a', 'copy']
-								: []),
+								: []))),
 					...(bitrateChanged && volume !== 0 ? ['-b:a', `${bitrate}k`] : []),
+					...(complexFilter ? ['-filter_complex', complexFilter, ...(complexVideoFilterPart.length >= 1 ? ['-map', '[v]'] : ['-map', '0:v']), ...(complexAudioFilterPart.length >= 1 ? ['-map', '[a]'] : ['-map', '0:a'])] : []),
 					...(
-						willCrop ? ['-vf', `crop=${cropWidth}:${cropHeight}:${cropX}:${cropY}`] :
-						extension === 'webm' && outExt === 'mp4' || compressionLevel !== 0
-						? [] : ['-c:v', 'copy']
+						(complexVideoFilterPart.length >= 1) ? [] :
+						(complexVideoFilterPart.length >= 1 || (extension === 'webm' && outExt === 'mp4' || compressionLevel !== 0)
+						? [] : ['-c:v', 'copy'])
 					),
 					...(compressionArgs[compressionLevel]),
 					`out.${outExt}`
-				]);
+				];
+				console.info("Running FFmpeg command");
+				await runFFmpeg(ffCommand);
+			}
 
 			processState = ProcessingState.READING;
 			const data = await ffmpeg.readFile(`out.${outExt}`);
-			const blob = new Blob([(data as Uint8Array).buffer]);
+			const blob = new Blob([(data as Uint8Array).buffer as BlobPart]);
 			const downloadURL = URL.createObjectURL(blob);
 			console.log(' ---- FINISHED ---- ');
 
@@ -201,6 +257,10 @@
 	let timelineWidth = 0;
 	let hoveredTime = -1;
 
+	// Resize variables
+	let resizeWidth = 0;
+	let resizeHeight = 0;
+
 	// Cropping variables
 	let cropX = 0;
 	let cropY = 0;
@@ -225,6 +285,9 @@
 	let trimReencoding = false;
 	let compressionLevel = 0;
 	let bitrate = 0;
+	let speedFactor = 1;
+	let keepPitch: boolean = false;
+	let semitoneFactor: number = 0;
 
 	const editorComponents: Record<
 		string,
@@ -269,6 +332,14 @@
 				showCropHandles = false;
 			}
 		},
+		resize: {
+			name: 'Resize',
+			icon: resizeIcon,
+			onClose() {
+				resizeHeight = 0;
+				resizeWidth = 0;
+			}
+		},
 		volume: {
 			name: 'Volume',
 			icon: volumeIcon,
@@ -292,13 +363,21 @@
 				compressionLevel = 0;
 			}
 		},
+		speed: {
+			name: 'Speed/Pitch',
+			icon: speedIcon,
+			onClose() {
+				speedFactor = 0;
+			}
+		},
 		bitrate: {
 			name: 'Bitrate',
 			icon: bitrateIcon,
 			onClose() {
 				bitrate = 0;
 			}
-		}
+		},
+
 	};
 
 	function onKeyPress(e: KeyboardEvent) {
@@ -641,6 +720,14 @@
 				on:setmode={(e) => (volumeMode = e.detail)}
 				on:setloudnorm={(e) => (loudnormArgs = e.detail)}
 			/>
+		{:else if tab === 'resize'}
+			<Resize
+				{resizeWidth} {resizeHeight}
+				on:set={(e) => {
+					if (e.detail.w !== undefined) resizeWidth = e.detail.w;
+					if (e.detail.h !== undefined) resizeHeight = e.detail.h;
+				}}
+			/>
 		{:else if tab === 'convert'}
 			<Convert {toExtension} {extension} on:set={(e) => (toExtension = e.detail)} />
 		{:else if tab === 'compress'}
@@ -659,6 +746,15 @@
 					if (e.detail.h !== undefined) cropHeight = e.detail.h;
 				}}
 			/>
+		{:else if tab === 'speed'}
+			<Speed
+				{semitoneFactor} {keepPitch} {speedFactor}
+				on:set={(e) => {
+					if (e.detail.s !== undefined) speedFactor = e.detail.s;
+					if (e.detail.t !== undefined) semitoneFactor = e.detail.t;
+				}}
+				on:setKeepPitch={(e) => keepPitch = e.detail}
+			/>
 		{/if}
 	</EditorTabs>
 </section>
@@ -668,5 +764,6 @@
 	open={modalOpen}
 	{processState}
 	{resultInfo}
+	{speedFactor}
 	on:close={() => (modalOpen = false)}
 />
